organize all files
