+++ tests/pretests/checkboxes.py
