+++ tests/pretests/entry_boxes_on_canvas.py
