+++ tests/pretests/entry_boxes_on_canvas_main.py
