+++ tests/pretests/new_windows_in_tkinter.py
