+++ tests/pretests/radio_buttons_2.py
