+++ tests/pretests/slider.py
