+++ tests/pretests/slider2.py
