+++ tests/pretests/slider3.py
