+++ tests/pretests/weather_imgs_format.py
